Add typing process and reset api instance

+++ src/main.js
@@ -15,6 +15,27 @@ const bot = new Telegraf(config.get('TELEGRAM_TOKEN'))
 
 bot.use(session())
 
+bot.command('system', async (ctx) => {
+  console.log('Command: system');
+
+  const hasAccess = await checkAccess(ctx, 'system')
+
+  if (!hasAccess) {
+    return
+  }
+
+  
+  ctx.session ??= { ...INITIAL_SESSION }
+  
+  await ctx.sendChatAction('typing')
+  const userText = ctx.message.text.replace('/system ', '')
+  ctx.session.messages.push({ role: openai.roles.System, content: userText })
+  const gptResponse = await openai.sendMessages(ctx.session.messages)
+  ctx.session.messages.push({ role: openai.roles.Assistant, content: gptResponse.content })
+  
+  await ctx.reply(code(`Системное сообщение передано: ${userText}`))
+})
+
 bot.command('new', async (ctx) => {
   console.log('Command: new');
 
@@ -24,6 +45,7 @@ bot.command('new', async (ctx) => {
     return
   }
 
+  openai.createNewApiInstance()
   ctx.session = { ...INITIAL_SESSION }
   await ctx.reply(code('Сессия сброшена'))
 })
@@ -40,26 +62,33 @@ bot.on(message('voice'), async ctx => {
   ctx.session ??= { ...INITIAL_SESSION }
 
   try {
-    await ctx.reply(code('processing...'))
+    await ctx.reply(code('start processing'))
+    await ctx.sendChatAction('typing')
 
     const voiceLink = await ctx.telegram.getFileLink(ctx.message.voice.file_id)
     const userId = ctx.message.from.id
 
     const oggFile = await ogg.create(voiceLink.href, userId)
     const mp3File = await ogg.toMp3(oggFile, userId)
-    const userText = await openai.transcription(mp3File)
+    const userText = await openai.getTranscription(mp3File)
     
     console.log(`User input(${ctx.message.from.username}): ${userText}`);
     await ctx.reply(code(`ваш запрос: ${userText}`))
+    await ctx.sendChatAction('typing')
 
     ctx.session.messages.push({ role: openai.roles.User, content: userText })
 
-    const responseText = await openai.chat(ctx.session.messages)
-    console.log(`GPT output(${ctx.message.from.username}): ${responseText}`);
+    const gptResponse = await openai.sendMessages(ctx.session.messages)
+    console.log(`GPT output(${ctx.message.from.username}): ${gptResponse.content}`);
 
-    ctx.session.messages.push({ role: openai.roles.Assistant, content: responseText })
+    ctx.session.messages.push({ role: openai.roles.Assistant, content: gptResponse.content })
     
-    await ctx.reply(responseText)
+    if (!gptResponse.content) {
+      await ctx.reply(code('GPT API is not responding. Try to reset session by /new command'))
+      return
+    }
+    
+    await ctx.reply(gptResponse.content)
     await removeFile(oggFile)
   } catch (error) {
     console.log('Voice message error:', error.message)
@@ -79,19 +108,25 @@ bot.on(message('text'), async ctx => {
   ctx.session ??= { ...INITIAL_SESSION }
 
   try {
-    await ctx.reply(code('processing...'))
+    await ctx.reply(code('start processing...'))
+    await ctx.sendChatAction('typing')
 
     const userText = ctx.message.text
     console.log(`User input(${ctx.message.from.username}): ${userText}`);
 
     ctx.session.messages.push({ role: openai.roles.User, content: userText })
 
-    const responseText = await openai.chat(ctx.session.messages)
-    console.log(`GPT output(${ctx.message.from.username}): ${responseText}`);
+    const gptResponse = await openai.sendMessages(ctx.session.messages)
+    console.log(`GPT output(${ctx.message.from.username}): ${gptResponse.content}`);
 
-    ctx.session.messages.push({ role: openai.roles.Assistant, content: responseText })
+    ctx.session.messages.push({ role: openai.roles.Assistant, content: gptResponse.content })
     
-    await ctx.reply(responseText)
+    if (!gptResponse.content) {
+      await ctx.reply(code('GPT API is not responding. Try to reset session by /new command'))
+      return
+    }
+    
+    await ctx.reply(gptResponse.content)
   } catch (error) {
     console.log('Text message error:', error.message)
     await ctx.reply(`Error: ${error.message}`)
